# merchant/src/main/resources/docs/swagger.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "Merchant API",
    "description": "UI and API endpoints for merchant microservice",
    "version": "1.0"
  },
  "servers": [
    {
      "url": "http://localhost:8081",
      "description": "Generated server url"
    }
  ],
  "security": [
    {
      "bearerAuth": []
    }
  ],
  "paths": {
    "/ui/merchants/{id}": {
      "get": {
        "tags": [
          "merchant-ui-controller"
        ],
        "summary": "Get a merchant by id",
        "description": "Returns a merchant data as per the id.",
        "operationId": "getMerchant",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "404": {
            "description": "The merchant was not found",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/MerchantViewDTO"
                }
              }
            }
          },
          "200": {
            "description": "Successfully retrieved",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/MerchantViewDTO"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "merchant-ui-controller"
        ],
        "summary": "Update a merchant data by id",
        "description": "Update a merchant. Returns an updated merchant data.",
        "operationId": "updateMerchant",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MerchantEditDTO"
              }
            }
          },
          "required": true
        },
        "responses": {
          "500": {
            "description": "Server error - check the server log",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/MerchantViewDTO"
                }
              }
            }
          },
          "200": {
            "description": "Successfully retrieved",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/MerchantViewDTO"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "merchant-ui-controller"
        ],
        "summary": "Remove a merchant by user auth ID",
        "description": "Delete a merchant based on the provided user ID.",
        "operationId": "removeMerchantById",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successfully removed the merchant",
            "content": {
              "*/*": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "string"
                  }
                }
              }
            }
          },
          "500": {
            "description": "Server error - check the server log",
            "content": {
              "*/*": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "string"
                  }
                }
              }
            }
          }
        }
      }
    },
...
    "/api/merchants/is-active": {
      "get": {
        "tags": [
          "merchant-api-controller"
        ],
        "summary": "Get a merchant status by uuid",
        "description": "Returns a merchant status data as per the uuid.",
        "operationId": "getIsMerchantActive",
        "responses": {
          "404": {
            "description": "The merchant was not found",
            "content": {
              "*/*": {
                "schema": {
                  "type": "boolean"
                }
              }
            }
          },
          "200": {
            "description": "Successfully retrieved",
            "content": {
              "*/*": {
                "schema": {
                  "type": "boolean"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "MerchantEditDTO": {
        "required": [
          "description",
          "email",
          "name",
          "status"
        ],
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "example": "ABC Store"
          },
          "description": {
            "type": "string",
            "example": "ABC Store in City 17"
          },
          "email": {
            "type": "string",
            "example": "company@mail.io"
          },
          "status": {
            "type": "string",
            "example": "ACTIVE",
            "enum": [
              "ACTIVE",
              "INACTIVE"
            ]
          }
        }
      },
      "MerchantViewDTO": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "email": {
            "type": "string"
          },
          "status": {
            "type": "string",
            "enum": [
              "ACTIVE",
              "INACTIVE"
            ]
          },
          "totalTransactionSum": {
            "type": "number"
          }
        }
      },
...
    },
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  }
}